+++ frontend/vue/components/Stations/Profile/backendPanelProps.js
@@ -1,13 +1,46 @@
 export default {
-    numSongs: Number,
-    numPlaylists: Number,
-    backendType: String,
-    hasStarted: Boolean,
-    userCanManageBroadcasting: Boolean,
-    userCanManageMedia: Boolean,
-    manageMediaUri: String,
-    managePlaylistsUri: String,
-    backendRestartUri: String,
-    backendStartUri: String,
-    backendStopUri: String
+    numSongs: {
+        type: Number,
+        required: true
+    },
+    numPlaylists: {
+        type: Number,
+        required: true
+    },
+    backendType: {
+        type: String,
+        required: true
+    },
+    hasStarted: {
+        type: Boolean,
+        required: true
+    },
+    userCanManageBroadcasting: {
+        type: Boolean,
+        required: true
+    },
+    userCanManageMedia: {
+        type: Boolean,
+        required: true
+    },
+    manageMediaUri: {
+        type: String,
+        required: true
+    },
+    managePlaylistsUri: {
+        type: String,
+        required: true
+    },
+    backendRestartUri: {
+        type: String,
+        required: true
+    },
+    backendStartUri: {
+        type: String,
+        required: true
+    },
+    backendStopUri: {
+        type: String,
+        required: true
+    }
 };