+++ frontend/vue/components/Stations/Profile/headerPanelProps.js
@@ -1,6 +1,18 @@
 export default {
-    stationName: String,
-    stationDescription: String,
-    userCanManageProfile: Boolean,
-    manageProfileUri: String
+    stationName: {
+        type: String,
+        required: true
+    },
+    stationDescription: {
+        type: String,
+        required: true
+    },
+    userCanManageProfile: {
+        type: Boolean,
+        required: true
+    },
+    manageProfileUri: {
+        type: String,
+        required: true
+    }
 }