+++ frontend/vue/components/Stations/Profile/publicPagesPanelProps.js
@@ -1,15 +1,58 @@
 export default {
-    stationSupportsStreamers: Boolean,
-    stationSupportsRequests: Boolean,
-    enablePublicPage: Boolean,
-    enableStreamers: Boolean,
-    enableOnDemand: Boolean,
-    enableRequests: Boolean,
-    userCanManageProfile: Boolean,
-    publicPageUri: String,
-    publicWebDjUri: String,
-    publicOnDemandUri: String,
-    publicPodcastsUri: String,
-    publicScheduleUri: String,
-    togglePublicPageUri: String
+    stationSupportsStreamers: {
+        type: Boolean,
+        required: true
+    },
+    stationSupportsRequests: {
+        type: Boolean,
+        required: true
+    },
+    enablePublicPage: {
+        type: Boolean,
+        required: true
+    },
+    enableStreamers: {
+        type: Boolean,
+        required: true
+    },
+    enableOnDemand: {
+        type: Boolean,
+        required: true
+    },
+    enableRequests: {
+        type: Boolean,
+        required: true
+    },
+    userCanManageProfile: {
+        type: Boolean,
+        required: true
+    },
+    publicPageUri: {
+        type: String,
+        required: true
+    },
+    publicWebDjUri: {
+        type: String,
+        required: true
+    },
+    publicOnDemandUri: {
+        type: String,
+        required: true
+    },
+    publicPodcastsUri: {
+        type: String,
+        required: true
+    },
+    publicScheduleUri: {
+        type: String,
+        required: true
+    },
+    togglePublicPageUri: {
+        type: String,
+        required: true
+    },
+    brandingUri: {
+        type: String,
+        required: true
+    },
 }