+++ frontend/vue/components/Stations/Profile/streamersPanelProps.js
@@ -1,7 +1,22 @@
 export default {
-    enableStreamers: Boolean,
-    userCanManageProfile: Boolean,
-    userCanManageStreamers: Boolean,
-    streamersViewUri: String,
-    streamersToggleUri: String
+    enableStreamers: {
+        type: Boolean,
+        required: true
+    },
+    userCanManageProfile: {
+        type: Boolean,
+        required: true
+    },
+    userCanManageStreamers: {
+        type: Boolean,
+        required: true
+    },
+    streamersViewUri: {
+        type: String,
+        required: true
+    },
+    streamersToggleUri: {
+        type: String,
+        required: true
+    }
 }