+++ frontend/vue/pages/Stations/Branding.js
@@ -1,0 +1,7 @@
+import initBase from '~/base.js';
+
+import '~/vendor/fancybox';
+
+import StationsBranding from '~/components/Stations/Branding.vue';
+
+export default initBase(StationsBranding);