+++ src/Assets/AbstractCustomAsset.php
@@ -4,14 +4,17 @@ declare(strict_types=1);
 
 namespace App\Assets;
 
+use App\Entity\Station;
 use App\Environment;
 use GuzzleHttp\Psr7\Uri;
 use Psr\Http\Message\UriInterface;
+use Symfony\Component\Filesystem\Filesystem;
 
 abstract class AbstractCustomAsset implements CustomAssetInterface
 {
     public function __construct(
-        protected readonly Environment $environment
+        protected readonly Environment $environment,
+        protected readonly ?Station $station = null
     ) {
     }
 
@@ -22,7 +25,18 @@ abstract class AbstractCustomAsset implements CustomAssetInterface
     public function getPath(): string
     {
         $pattern = sprintf($this->getPattern(), '');
-        return $this->environment->getUploadsDirectory() . '/' . $pattern;
+        return $this->getBasePath() . '/' . $pattern;
+    }
+
+    protected function getBasePath(): string
+    {
+        $basePath = $this->environment->getUploadsDirectory();
+
+        if (null !== $this->station) {
+            $basePath .= '/' . $this->station->getShortName();
+        }
+
+        return $basePath;
     }
 
     public function getUrl(): string
@@ -32,13 +46,24 @@ abstract class AbstractCustomAsset implements CustomAssetInterface
             $pattern = $this->getPattern();
             $mtime = filemtime($path);
 
-            return $this->environment->getAssetUrl() . self::UPLOADS_URL_PREFIX . '/' . sprintf(
+            return $this->getBaseUrl() . '/' . sprintf(
                 $pattern,
                 '.' . $mtime
             );
         }
 
         return $this->getDefaultUrl();
+    }
+
+    protected function getBaseUrl(): string
+    {
+        $baseUrl = $this->environment->getAssetUrl() . self::UPLOADS_URL_PREFIX;
+
+        if (null !== $this->station) {
+            $baseUrl .= '/' . $this->station->getShortName();
+        }
+
+        return $baseUrl;
     }
 
     public function getUri(): UriInterface
@@ -54,5 +79,10 @@ abstract class AbstractCustomAsset implements CustomAssetInterface
     public function delete(): void
     {
         @unlink($this->getPath());
+    }
+
+    protected function ensureDirectoryExists(string $path): void
+    {
+        (new Filesystem())->mkdir(dirname($path));
     }
 }