+++ src/Assets/BrowserIconCustomAsset.php
@@ -41,7 +41,7 @@ final class BrowserIconCustomAsset extends AbstractCustomAsset
     public function upload(Image $image): void
     {
         $uploadsDir = $this->environment->getUploadsDirectory() . '/browser_icon';
-        (new Filesystem())->mkdir($uploadsDir);
+        $this->ensureDirectoryExists($uploadsDir);
 
         $newImage = clone $image;
         $newImage->resize(256, 256);