+++ src/Entity/Repository/StationRepository.php
@@ -181,6 +181,11 @@ final class StationRepository extends Repository
     public function getDefaultAlbumArtUrl(?Entity\Station $station = null): UriInterface
     {
         if (null !== $station) {
+            $stationAlbumArt = new AlbumArtCustomAsset($this->environment, $station);
+            if ($stationAlbumArt->isUploaded()) {
+                return $stationAlbumArt->getUri();
+            }
+
             $stationCustomUri = $station->getBrandingConfig()->getDefaultAlbumArtUrlAsUri();
             if (null !== $stationCustomUri) {
                 return $stationCustomUri;