+++ src/Entity/StationBrandingConfiguration.php
@@ -29,4 +29,28 @@ class StationBrandingConfiguration extends AbstractStationConfiguration
     {
         $this->set(self::DEFAULT_ALBUM_ART_URL, $default_album_art_url);
     }
+
+    public const PUBLIC_CUSTOM_CSS = 'public_custom_css';
+
+    public function getPublicCustomCss(): ?string
+    {
+        return $this->get(self::PUBLIC_CUSTOM_CSS);
+    }
+
+    public function setPublicCustomCss(?string $css): void
+    {
+        $this->set(self::PUBLIC_CUSTOM_CSS, $css);
+    }
+
+    public const PUBLIC_CUSTOM_JS = 'public_custom_js';
+
+    public function getPublicCustomJs(): ?string
+    {
+        return $this->get(self::PUBLIC_CUSTOM_JS);
+    }
+
+    public function setPublicCustomJs(?string $js): void
+    {
+        $this->set(self::PUBLIC_CUSTOM_JS, $js);
+    }
 }